Added method for updating status based on last update ts
The gossiper starts a sub-routine which periodically goes over the node
data, and for all nodes for which the last update timestamp was older
than a configured value marks their status as down.

Note that this status will be over-written if we get a newer set of
information from a peer.